// api/src/db/schema.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

export interface Day {
  complete: number;
  date: string;
  raw: string;
  workout_calories: number | null;
  workout_note: string | null;
}

export interface DayEntry {
  calories: number;
  calories_100g: number | null;
  date: string;
  food_id: string | null;
  foodname: string | null;
  grams: number | null;
  id: string;
  override_calories: number | null;
  override_calories_100g: number | null;
  raw: string;
  sort: number;
}

export interface Food {
  added_on: string;
  calories: number;
  carbs: number;
  fats: number;
  fiber: number;
  id: string;
  name: string;
  protein: number;
}

export interface Session {
  expires_on: number;
  id: string;
  user_id: string;
}

export interface User {
  created_on: number;
  id: string;
  password_hash: string;
  updated_on: number;
  username: string;
}

export interface DB {
  day: Day;
  day_entry: DayEntry;
  food: Food;
  session: Session;
  user: User;
}